Deprecate fastf1.utils submodule and move implementations to internals
Move the implementations of recursive_dict_get, to_datetime, and
to_timedelta to fastf1.internals._utils, which is the appropriate
location for internal utilities. Update all internal imports in
events.py, _api.py, core.py, and livetiming/data.py to use the new
internal module directly.

The public API in fastf1.utils is now deprecated wrappers that:
1. Emit a DeprecationWarning when called
2. Delegate to the internal implementations
3. Maintain backward compatibility for external callers

Fixes #884
