+++ fastf1/internals/_utils.py
@@ -1,0 +1,124 @@
+"""Internal utility functions.
+
+These functions are not part of the public API and may be changed
+or removed at any time without notice.
+"""
+from functools import reduce
+
+import datetime
+
+
+def recursive_dict_get(d: dict, *keys: str, default_none: bool = False):
+    """Recursive dict get. Can take an arbitrary number of keys and returns an
+    empty dict if any key does not exist.
+    https://stackoverflow.com/a/28225747"""
+    ret = reduce(lambda c, k: c.get(k, {}), keys, d)
+    if default_none and ret == {}:
+        return None
+    return ret
+
+
+def to_timedelta(x: str | datetime.timedelta) \
+        -> datetime.timedelta | None:
+    """Fast timedelta object creation from a time string.
+
+    Permissible string formats:
+
+        For example: `13:24:46.320215` with:
+
+            - optional hours and minutes
+            - optional microseconds and milliseconds with
+              arbitrary precision (1 to 6 digits)
+
+        Examples of valid formats:
+
+            - `24.3564` (seconds + milli/microseconds)
+            - `36:54` (minutes + seconds)
+            - `8:45:46` (hours, minutes, seconds)
+
+    Args:
+        x: timestamp
+    """
+    # this is faster than using pd.timedelta on a string
+    if isinstance(x, str) and len(x):
+        try:
+            hours, minutes = 0, 0
+            if len(hms := x.split(':')) == 3:
+                hours, minutes, seconds = hms
+            elif len(hms) == 2:
+                minutes, seconds = hms
+            else:
+                seconds = hms[0]
+
+            if '.' in seconds:
+                seconds, msus = seconds.split('.')
+                if len(msus) < 6:
+                    msus = msus + '0' * (6 - len(msus))
+                elif len(msus) > 6:
+                    msus = msus[0:6]
+            else:
+                msus = 0
+
+            return datetime.timedelta(
+                hours=int(hours), minutes=int(minutes),
+                seconds=int(seconds), microseconds=int(msus)
+            )
+
+        except Exception as exc:
+            return None
+
+    elif isinstance(x, datetime.timedelta):
+        return x
+
+    else:
+        return None
+
+
+def to_datetime(x: str | datetime.datetime) \
+        -> datetime.datetime | None:
+    """Fast datetime object creation from a date string.
+
+    Permissible string formats:
+
+        For example '2020-12-13T13:27:15.320000Z' with:
+
+            - optional milliseconds and microseconds with
+              arbitrary precision (1 to 6 digits)
+            - with optional trailing letter 'Z'
+
+        Examples of valid formats:
+
+            - `2020-12-13T13:27:15.320000`
+            - `2020-12-13T13:27:15.32Z`
+            - `2020-12-13T13:27:15`
+
+    Args:
+        x: timestamp
+    """
+    if isinstance(x, str) and x:
+        try:
+            date, time = x.strip('Z').split('T')
+            year, month, day = date.split('-')
+            hours, minutes, seconds = time.split(':')
+            if '.' in seconds:
+                seconds, msus = seconds.split('.')
+                if len(msus) < 6:
+                    msus = msus + '0' * (6 - len(msus))
+                elif len(msus) > 6:
+                    msus = msus[0:6]
+            else:
+                msus = 0
+
+            return datetime.datetime(
+                int(year), int(month), int(day), int(hours),
+                int(minutes), int(seconds), int(msus)
+            )
+
+        except Exception:
+            return None
+
+    elif isinstance(x, datetime.datetime):
+        return x
+
+    else:
+        return None
