+++ fastf1/utils.py
@@ -1,16 +1,10 @@
 """This is a collection of various functions."""
-import datetime
 import warnings
-from functools import reduce
 
 import numpy as np
 import pandas as pd
 
 import fastf1
-from fastf1.logger import get_logger
-
-
-_logger = get_logger(__name__)
 
 
 def delta_time(
@@ -109,120 +103,71 @@ def delta_time(
 
 
 def recursive_dict_get(d: dict, *keys: str, default_none: bool = False):
-    """Recursive dict get. Can take an arbitrary number of keys and returns an
-    empty dict if any key does not exist.
-    https://stackoverflow.com/a/28225747"""
-    ret = reduce(lambda c, k: c.get(k, {}), keys, d)
-    if default_none and ret == {}:
-        return None
-    return ret
+    """Recursive dict get.
 
-
-def to_timedelta(x: str | datetime.timedelta) \
-        -> datetime.timedelta | None:
-    """Fast timedelta object creation from a time string
-
-    Permissible string formats:
-
-        For example: `13:24:46.320215` with:
-
-            - optional hours and minutes
-            - optional microseconds and milliseconds with
-              arbitrary precision (1 to 6 digits)
-
-        Examples of valid formats:
-
-            - `24.3564` (seconds + milli/microseconds)
-            - `36:54` (minutes + seconds)
-            - `8:45:46` (hours, minutes, seconds)
+    .. deprecated:: 3.2.0
+        This function was never meant to be part of the public API and
+        has been moved to ``fastf1.internals._utils``. Use that instead.
 
     Args:
-        x: timestamp
+        d: dictionary
+        keys: variable number of keys
+        default_none: return None instead of empty dict on missing key
+
+    Returns:
+        The value at the nested key path, or None/empty dict on missing keys.
     """
-    # this is faster than using pd.timedelta on a string
-    if isinstance(x, str) and len(x):
-        try:
-            hours, minutes = 0, 0
-            if len(hms := x.split(':')) == 3:
-                hours, minutes, seconds = hms
-            elif len(hms) == 2:
-                minutes, seconds = hms
-            else:
-                seconds = hms[0]
-
-            if '.' in seconds:
-                seconds, msus = seconds.split('.')
-                if len(msus) < 6:
-                    msus = msus + '0' * (6 - len(msus))
-                elif len(msus) > 6:
-                    msus = msus[0:6]
-            else:
-                msus = 0
-
-            return datetime.timedelta(
-                hours=int(hours), minutes=int(minutes),
-                seconds=int(seconds), microseconds=int(msus)
-            )
-
-        except Exception as exc:
-            _logger.debug(f"Failed to parse timedelta string '{x}'",
-                          exc_info=exc)
-            return None
-
-    elif isinstance(x, datetime.timedelta):
-        return x
-
-    else:
-        return None
+    warnings.warn(
+        "`fastf1.utils.recursive_dict_get` is deprecated and will be removed. "
+        "Use `fastf1.internals._utils.recursive_dict_get` instead.",
+        DeprecationWarning,
+        stacklevel=2
+    )
+    from fastf1.internals._utils import recursive_dict_get as _rdg
+    return _rdg(d, *keys, default_none=default_none)
 
 
-def to_datetime(x: str | datetime.datetime) \
-        -> datetime.datetime | None:
+def to_timedelta(x):
+    """Fast timedelta object creation from a time string.
+
+    .. deprecated:: 3.2.0
+        This function was never meant to be part of the public API and
+        has been moved to ``fastf1.internals._utils``. Use that instead.
+
+    Args:
+        x: timestamp string or datetime.timedelta
+
+    Returns:
+        datetime.timedelta or None
+    """
+    warnings.warn(
+        "`fastf1.utils.to_timedelta` is deprecated and will be removed. "
+        "Use `fastf1.internals._utils.to_timedelta` instead.",
+        DeprecationWarning,
+        stacklevel=2
+    )
+    from fastf1.internals._utils import to_timedelta as _ttd
+    return _ttd(x)
+
+
+def to_datetime(x):
     """Fast datetime object creation from a date string.
 
-    Permissible string formats:
-
-        For example '2020-12-13T13:27:15.320000Z' with:
-
-            - optional milliseconds and microseconds with
-              arbitrary precision (1 to 6 digits)
-            - with optional trailing letter 'Z'
-
-        Examples of valid formats:
-
-            - `2020-12-13T13:27:15.320000`
-            - `2020-12-13T13:27:15.32Z`
-            - `2020-12-13T13:27:15`
+    .. deprecated:: 3.2.0
+        This function was never meant to be part of the public API and
+        has been moved to ``fastf1.internals._utils``. Use that instead.
 
     Args:
-        x: timestamp
+        x: timestamp string or datetime.datetime
+
+    Returns:
+        datetime.datetime or None
     """
-    if isinstance(x, str) and x:
-        try:
-            date, time = x.strip('Z').split('T')
-            year, month, day = date.split('-')
-            hours, minutes, seconds = time.split(':')
-            if '.' in seconds:
-                seconds, msus = seconds.split('.')
-                if len(msus) < 6:
-                    msus = msus + '0' * (6 - len(msus))
-                elif len(msus) > 6:
-                    msus = msus[0:6]
-            else:
-                msus = 0
-
-            return datetime.datetime(
-                int(year), int(month), int(day), int(hours),
-                int(minutes), int(seconds), int(msus)
-            )
-
-        except Exception as exc:
-            _logger.debug(f"Failed to parse datetime string '{x}'",
-                          exc_info=exc)
-            return None
-
-    elif isinstance(x, datetime.datetime):
-        return x
-
-    else:
-        return None
+    warnings.warn(
+        "`fastf1.utils.to_datetime` is deprecated and will be removed. "
+        "Use `fastf1.internals._utils.to_datetime` instead.",
+        DeprecationWarning,
+        stacklevel=2
+    )
+    from fastf1.internals._utils import to_datetime as _tdt
+    return _tdt(x)
